# SupplierService/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Supplier Microservice API",
        "contact": {}
    },
    "paths": {
        "/deliveries/show": {
            "get": {
                "description": "Get the list of all deliveries",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deliveries"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repository.Delivery"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/suppliers/add": {
            "post": {
                "description": "Add a new supplier to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "parameters": [
                    {
                        "description": "Supplier to add",
                        "name": "supplier",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.AddSupplierRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/suppliers/delete": {
            "delete": {
                "description": "Delete a supplier from the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "parameters": [
                    {
                        "description": "Supplier to delete",
                        "name": "supplier",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.DeleteSupplierRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/suppliers/show": {
            "get": {
                "description": "Get the list of all suppliers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repository.Supplier"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "main.AddSupplierRequest": {
            "type": "object",
            "properties": {
                "company_name": {
                    "type": "string"
                },
                "days_to_ship": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "main.DeleteSupplierRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "repository.Delivery": {
            "type": "object",
            "properties": {
                "arrival": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "repository.Supplier": {
            "type": "object",
            "properties": {
                "company_name": {
                    "type": "string"
                },
                "days_to_ship": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                }
            }
        }
    }
}